asm,btf: added *btf.Func reference to per-instruction metadata
This commit adds a *btf.Func reference to the per-instruction metadata.
We want to do this so in a future commit we can use the referenced func
to generate .BTF.ext blobs from the per instruction-metadata.

Just adding the function metadata wouldn't have worked since it would
have caused an import cycle. The btf package utilized the
`asm.InstructionSize` const and referenced the asm package to apply
CO-RE fixups. To remove these dependencies, the const was cloned to the
btf package since it is not expected to change. The fixup functions were
moved from the btf package to the asm package which inverts the
dependency thus making the btf package no longer dependant of asm,
rather asm is now dependant on btf.