btf: Move type parsing into decoder structure
This commit moves the type parsing into the decoder structure. This
decoder allows for the parsing and inflation of single BTF types and
its dependencies by ID, while still supporting full parsing of BTF.

This is in preperation for lazy decoding, this way we can share the
parsing and inflation logic between the two strategies.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>